requirements.txt
py-cord==2.6.1
anthropic>=0.25.0
httpx>=0.27.0
pandas>=2.2.0
fredapi>=0.5.1
python-dotenv>=1.0.0
aiohttp>=3.9.0
PyMuPDF>=1.24.0
apscheduler>=3.10.0
